+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## 1.1.6 - 2017-08-18
+## 1.1.7 - 2017-08-19
+### Added
+- New **Button** prop, `relax`.
+- New **Divider** color, `light`.
+- New global text color, `color-success`
+
+## 1.1.6 - 2017-08-19
 ### Added
 - **Dropdown** (not selections) menu's now detect if they are in the viewport.
 - Added a `browserDetect` function.